+++ package.json
@@ -38,6 +38,7 @@
     "@mui/x-date-pickers": "^6.5.0",
     "@next/font": "^13.0.5",
     "@paypal/react-paypal-js": "^7.8.1",
+    "@ramonak/react-progress-bar": "^5.0.3",
     "@react-pdf/renderer": "^3.1.3",
     "@sentry/nextjs": "7.21.1",
     "@stripe/react-stripe-js": "^1.16.1",
