+++ public/locales/bg/campaigns.json
@@ -87,7 +87,6 @@
     "amount": "Необходима сума",
     "type": "Тип кампания",
     "state": "Статус",
-    "reached": "Събрани:",
     "target": "Цел:",
     "from": "от",
     "noCommissionInfo": "Подкрепи.бг работи с 0% комисиона. Заплащат се единствено банкови такси, които изрично се упоменават преди да направите дарението си.",
