+++ src/components/client/campaigns/InlineDonation.tsx
@@ -23,7 +23,7 @@ import theme from 'common/theme'
 import { useCopyToClipboard } from 'common/util/useCopyToClipboard'
 import useMobile from 'common/hooks/useMobile'
 import LinkButton from '../../common/LinkButton'
-import CampaignProgress from './CampaignProgress'
+import CampaignProgress from './CampaignProgress/CampaignProgress'
 import DonorsAndDonations from './DonorsAndDonations'
 import CustomListItem from 'components/common/navigation/CustomListItem'
 import { socialMedia } from './helpers/socialMedia'
